changes added

+++ src/main/java/com/shreya/hibernate/repository/CartItemRepository.java
@@ -4,6 +4,6 @@ import com.shreya.hibernate.domain.CartItemDomain;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CartItemRepository extends JpaRepository<CartItemDomain, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItemDomain, Long> {
 
 }
